Add tests and reorganize directory
